requirements.txt
google-generativeai>=0.7.0
chromadb>=0.4.22
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
pdfplumber>=0.11.0
gradio>=4.31.0
torch>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
groq>=0.9.0
